Disable redeploy when no change is detected

+++ update-hosts-domains.sh
@@ -1,5 +1,11 @@
+cp hosts.js hosts.js.bkp
 python3 ./update-hosts-domain.py > hosts.js
-npm install beamup-cli -g
-npx beamup config a.baby-beamup.club $GITHUBUSERNAME
-npx beamup init lobo-guara-addon
-npx beamup deploy
+if cmp -s hosts.js hosts.js.bkp; then
+    echo "Hosts are up-to-date, no need for changes."
+else
+    npm install beamup-cli -g
+    npx beamup config a.baby-beamup.club $GITHUBUSERNAME
+    npx beamup init lobo-guara-addon
+    npx beamup deploy
+fi
+rm hosts.js.bkp
